+++ packages/vendure-plugin-invoices/test/dev-server.ts
@@ -45,10 +45,21 @@ require('dotenv').config();
             clientSecret: process.env.XERO_CLIENT_SECRET!,
             shippingAccountCode: '0103',
             salesAccountCode: '0102',
+            invoiceBrandingThemeId: '62f2bce1-32c4-4e8d-a9b1-87060fb7c791',
             getReference: () =>
               'THIS IS A TEST INVOICE, DONT APPROVE THIS PLEASE.',
             getVendureUrl: (order) =>
               `https://pinelab.studio/order/${order.code}`,
+            getDueDate: (ctx, order, invoice) => {
+              const payment = order.payments.find((p) => p.state === 'Settled');
+              if (payment?.method === 'purchase-order') {
+                const date = new Date();
+                date.setDate(date.getDate() + 30); //30 days later
+                return date;
+              } else {
+                return new Date();
+              }
+            },
           }),
         ],
       }),